@@ -3,8 +3,9 @@
     "version": "1.0.0",
     "private": true,
     "devDependencies": {
-        "babel-eslint": "^10.1.0",
-        "eslint": "^8.6.0",
+        "@babel/core": "^7.16.7",
+        "@babel/eslint-parser": "^7.16.5",
+        "eslint": "^8.7.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-react": "^7.28.0",
@@ -12,17 +13,17 @@
         "react-scripts": "^5.0.0"
     },
     "dependencies": {
-        "@material-ui/core": "^4.11.0",
-        "@material-ui/icons": "^4.9.1",
+        "@material-ui/core": "^4.12.3",
+        "@material-ui/icons": "^4.11.2",
         "@material-ui/lab": "^4.0.0-alpha.56",
         "axios": "^0.24.0",
         "dayjs": "^1.10.7",
-        "firebase": "^8.2.4",
+        "firebase": "^9.6.3",
         "prop-types": "^15.8.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-redux": "^7.2.6",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.2.1",
         "recharts": "^2.1.8",
         "redux": "^4.1.2",
         "redux-thunk": "^2.4.1",
