@@ -3,7 +3,7 @@ import Home from "Home";
 import Layout from "Layout";
 import PrivacyPolicy from "PrivacyPolicy";
 import React from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import ScrollToTop from "ScrollToTop";
 import TermsAndCondition from "TermsAndCondition";
 import { isDonateEnable } from "util/Util";
@@ -13,19 +13,17 @@ function App() {
         <Router>
             <ScrollToTop />
             <Layout>
-                <Switch>
-                    <Route path="/" exact component={Home} />
-                    <Route path="/termsAndCondition" exact component={TermsAndCondition} />
-                    <Route path="/privacyPolicy" exact component={PrivacyPolicy} />
+                <Routes>
+                    <Route path="/" element={<Home />} />
+                    <Route path="/termsAndCondition" element={<TermsAndCondition />} />
+                    <Route path="/privacyPolicy" element={<PrivacyPolicy />} />
                     {isDonateEnable() && (
                         <>
-                            <Route path="/donate/complete" exact>
-                                <Donation complete={true} />
-                            </Route>
-                            <Route path="/donate" component={Donation} exact />
+                            <Route path="/donate/complete" element={<Donation complete={true} />} />
+                            <Route path="/donate" element={<Donation />} />
                         </>
                     )}
-                </Switch>
+                </Routes>
             </Layout>
         </Router>
     );
